The code is put together now. Cleaned up as well. Needs a bit of commenting time-to-time, but works and is good. There is still a few things to do in the master.py -> Save the plot that is generated by the python file, Save the data that is generated by the predictions, Make the user-interface. So basically very few things are left - we can start working on the paper in January. NOTE: Deleted a bunch of code, that I found not useful, we can talk about what to do with them late - the deleted code is in droppedFunctions.py.
